\name{H2Result-methods}
\docType{methods}
\alias{H2Result-methods}
\alias{fetch,H2Result,numeric-method}
\title{Methods for the class H2Result in Package `H2' }
\description{
  Methods for the class `H2Result' in Package `H2'.

  \code{fetch} retrieves the content of the result set in the form of a
  data frame. 
 
  Other methods are inherited from the RJDBC package.

  }
\section{Methods}{
  \describe{
    \item{fetch}{\code{signature(res = "H2Result", ...)}}
}}
\keyword{methods}

