\name{H2Connection-methods}
\docType{methods}
\alias{H2Connection-methods}
\alias{dbGetQuery,H2Connection,character-method}
\alias{dbWriteTable,H2Connection-method}
\alias{dbDataType,H2Connection-method}
\alias{dbSendQuery,H2Connection,character-method}
\title{Methods for the class `H2Connect' in Package `H2'}
\description{
  Methods for the class `H2Connection' in Package `H2'.  
}
\section{Methods}{
  \describe{
    \item{dbWriteTable}{\code{signature(conn = "H2Connection", ...)}}

}}
\keyword{methods}

